app/requirements.txt
streamlit
langchain
llama-index
openai
chromadb
python-dotenv
pymupdf
